Seperate tests related to worker from test_freqtradebot
